Fix image compress

+++ BotDeScans.App/Services/ImageService.cs
@@ -3,76 +3,43 @@ using Microsoft.Extensions.Configuration;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Advanced;
 using SixLabors.ImageSharp.PixelFormats;
-using System.Runtime.InteropServices;
 namespace BotDeScans.App.Services;
 
 public class ImageService(IConfiguration configuration)
 {
-    public async Task CompressImagesAsync(string directory, CancellationToken cancellationToken = default)
+    public virtual async Task CompressImageAsync(
+        string filePath,
+        bool isGrayscale,
+        CancellationToken cancellationToken)
     {
-        var maxDegreeOfParallelism = configuration.GetValue<int?>("Compress:ParallelismDegree")
-            ?? Convert.ToInt32(Math.Ceiling((Environment.ProcessorCount * 0.75) * 2.0));
+        var (quality, minQuality) = GetImageQuality(isGrayscale);
+        var imageBytes = await TryCompressAsync(filePath, quality, minQuality);
+        if (imageBytes == null)
+            return;
 
-        var parallelOptions = new ParallelOptions
-        {
-            CancellationToken = cancellationToken,
-            MaxDegreeOfParallelism = maxDegreeOfParallelism
-        };
+        using var stream = File.Create(Path.ChangeExtension(filePath, ".png"));
+        await stream.WriteAsync(imageBytes.Value, cancellationToken);
 
-        await Parallel.ForEachAsync(
-            Directory.GetFiles(directory),
-            parallelOptions,
-            async (filePath, ct) =>
-        {
-            var (quality, minQuality) = GetImageQuality(filePath);
-
-            using var imageJob = new ImageJob();
-            await using var fileStream = File.OpenRead(filePath);
-            var imageJobResult = await imageJob
-                .Decode(source: BufferedStreamSource.UseEntireStreamAndDisposeWithSource(fileStream),
-                        commands: new DecodeCommands().SetIgnoreColorProfileErrors(true))
-                .EncodeToBytes(new PngQuantEncoder(quality, minQuality))
-                .Finish()
-                .InProcessAsync();
-
-            var bytes = imageJobResult.First?.TryGetBytes();
-            if (bytes != null)
-            {
-                var newFileName = $"{Path.GetFileNameWithoutExtension(filePath)}.png";
-                using var stream = File.Create(Path.Combine(directory, newFileName));
-                await stream.WriteAsync(bytes.Value, ct);
-            }
-
-            if (Path.GetExtension(filePath) != ".png")
-                File.Delete(filePath);
-        });
+        if (Path.GetExtension(filePath) != ".png")
+            File.Delete(filePath);
     }
 
-    public async Task<string> CreateBase64File(
+    public virtual async Task<string> CreateBase64String(
         string filePath,
         int maxWidth,
         int maxHeight,
+        bool isGrayscale,
         CancellationToken cancellationToken)
     {
-        var (quality, minQuality) = GetImageQuality(filePath);
+        var command = $"width={maxWidth}&height={maxHeight}&mode=max&scale=both";
+        var (quality, minQuality) = GetImageQuality(isGrayscale);
+        var imageBytes = await TryCompressAsync(filePath, quality, minQuality, command);
 
-        using var imageJob = new ImageJob();
-        var imageBytes = await File.ReadAllBytesAsync(filePath, cancellationToken);
-        var imageJobResult = await imageJob
-            .Decode(imageBytes)
-            .ResizerCommands($"width={maxWidth}&height={maxHeight}&mode=max&scale=both")
-            .EncodeToBytes(new PngQuantEncoder(quality, minQuality))
-            .Finish()
-            .InProcessAsync();
-
-        var convertedImageInBytes = imageJobResult.First?.TryGetBytes()
-            ?? throw new Exception("Unable to convert image");
-
-        return Convert.ToBase64String(convertedImageInBytes);
+        return Convert.ToBase64String(imageBytes ?? throw new Exception("Unable to convert image to base64"));
     }
 
-    private (int quality, int minQuality) GetImageQuality(string filePath)
-        => IsGrayscale(filePath, 10)
+    private (int quality, int minQuality) GetImageQuality(bool isGrayscale)
+        => isGrayscale
             ? (configuration.GetValue<int?>("Compress:Grayscale:Quality") ?? 50,
                configuration.GetValue<int?>("Compress:Grayscale:MinimumQuality") ?? 30)
             : (configuration.GetValue<int?>("Compress:Colorful:Quality") ?? 90,
@@ -83,30 +50,48 @@ public class ImageService(IConfiguration configuration)
     /// Determine if an image is greyscale
     /// </summary>
     /// <param name="filePath">The path to the image file.</param>
-    /// <param name="threshold"></param>
+    /// <param name="threshold">Color variation to ignore</param>
     /// <returns></returns>
-    private static bool IsGrayscale(string filePath, int threshold)
+    public virtual bool IsGrayscale(string filePath, int threshold = 10)
     {
         //Load image
-        using var image = Image.Load<Rgba32>(filePath);
-
-
+        using var fileStream = new FileStream(filePath, FileMode.Open, FileAccess.Read);
+        using var image = Image.Load<Rgba32>(fileStream);
         foreach (var row in image.GetPixelMemoryGroup())
-            foreach (var pixel in row.Span)
-            {
-                if (pixel.A == 0) //ignore fully transparent pixels 
-                    continue;
+        foreach (var pixel in row.Span)
+        {
+            if (pixel.A == 0) //ignore fully transparent pixels 
+                continue;
 
-                if (GetRgbDelta(pixel.R, pixel.G, pixel.B) > threshold)
-                    return false;
-            }
+            if (GetRgbDelta(pixel.R, pixel.G, pixel.B) > threshold)
+                return false;
+        }
+
         return true;
-
 
         static int GetRgbDelta(byte r, byte g, byte b)
             => Math.Abs(r - g) +
                Math.Abs(g - b) +
                Math.Abs(b - r);
 
+    }
+
+    private static async Task<ArraySegment<byte>?> TryCompressAsync(
+        string filePath,
+        int quality,
+        int minQuality,
+        string resizerCommands = "")
+    {
+        using var fileStream = File.OpenRead(filePath);
+        using var imageJob = new ImageJob();
+        var imageJobResult = await imageJob
+            .Decode(source: BufferedStreamSource.UseEntireStreamAndDisposeWithSource(fileStream),
+                    commands: new DecodeCommands().SetIgnoreColorProfileErrors(true))
+            .ResizerCommands(resizerCommands)
+            .EncodeToBytes(new PngQuantEncoder(quality, minQuality))
+            .Finish()
+            .InProcessAsync();
+
+        return imageJobResult.First?.TryGetBytes();
     }
 }
